code refactored

+++ mysql/functions.php
@@ -2,6 +2,34 @@
 
 include "db.php";
 
+
+function createRows() {
+	global $connection;
+
+	if(isset($_POST['submit'])) {
+
+
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+
+		$query = "INSERT INTO users(username, password)";
+		$query .= "VALUES('$username', '$password')";
+
+		$result = mysqli_query($connection, $query);
+
+		if(!$result) {
+			die('Query FAILED!' . mysqli_error());
+		} else {
+			echo "Record Created";
+		}	
+	}	
+}
+
+
+
+
+
+
 function showAllData() {
 	global $connection;
 
@@ -25,40 +53,48 @@ function showAllData() {
 
 
 function updateTable() {
-	global $connection;
+	if(isset($_POST['submit'])) {
+		global $connection;
 
-	$username = $_POST['username'];
-	$password = $_POST['password'];
-	$id = $_POST['id'];
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+		$id = $_POST['id'];
 
-	$query = "UPDATE users SET ";
-	$query .= "username = '$username', ";
-	$query .= "password = '$password' ";
-	$query .= "WHERE id = $id ";
+		$query = "UPDATE users SET ";
+		$query .= "username = '$username', ";
+		$query .= "password = '$password' ";
+		$query .= "WHERE id = $id ";
 
-	$result = mysqli_query($connection, $query);
+		$result = mysqli_query($connection, $query);
 
-	if(!$result) {
-		die("QUERY FAILED" . mysqli_error($connection));
+		if(!$result) {
+			die("QUERY FAILED" . mysqli_error($connection));
 
+		} else {
+			echo "Record Updated";
+		}
 	}
 }
 
 function deleteRows() {
-	global $connection;
+	if(isset($_POST['submit'])) {
+		global $connection;
 
-	$username = $_POST['username'];
-	$password = $_POST['password'];
-	$id = $_POST['id'];
+		$username = $_POST['username'];
+		$password = $_POST['password'];
+		$id = $_POST['id'];
 
-	$query = "DELETE FROM users ";
-	$query .= "WHERE id = $id ";
+		$query = "DELETE FROM users ";
+		$query .= "WHERE id = $id ";
 
-	$result = mysqli_query($connection, $query);
+		$result = mysqli_query($connection, $query);
 
-	if(!$result) {
-		die("QUERY FAILED" . mysqli_error($connection));
+		if(!$result) {
+			die("QUERY FAILED" . mysqli_error($connection));
 
+		} else {
+			echo "Record Deleted";
+		}
 	}
 }
